requirements.txt
Django~=3.1.3
django-widget-tweaks~=1.4.8  # フォームをきれいに表示する
django-allauth~=0.41.0  # Djangoのログイン機能(ユーザ認証)の実装
Pillow~=8.1.2  # pythonの画像処理ライブラリ
django-heroku~=0.3.1
gunicorn~=20.0.4